Updated functional tests to use pytest
