+++ .github/workflows/tests.yml
@@ -93,7 +93,7 @@ jobs:
         shell: python
 
       - name: Run tests
-        run: PYTHONPATH=$PWD/tests pytest -vv tests/ --cov=covalent_awslambda_plugin
+        run: PYTHONPATH=$PWD/tests pytest -m "not functional_tests" -vv tests/ --cov=covalent_awslambda_plugin
 
       - name: Generate coverage report
         run: coverage xml -o coverage.xml
