+++ pyproject.toml
@@ -37,3 +37,8 @@ multi_line_output = 3
 include_trailing_comma = true
 profile = 'black'
 skip_gitignore = true
+
+[tool.pytest.ini_options]
+markers = [
+    "functional_tests: marks tests that are to be run in the functional tests ci pipeline"
+]
