--- tests/functional_tests/basic_workflow.py
@@ -1,51 +1,0 @@
-# Copyright 2021 Agnostiq Inc.
-#
-# This file is part of Covalent.
-#
-# Licensed under the GNU Affero General Public License 3.0 (the "License").
-# A copy of the License may be obtained with this software package or at
-#
-#      https://www.gnu.org/licenses/agpl-3.0.en.html
-#
-# Use of this file is prohibited except in compliance with the License. Any
-# modifications or derivative works of this file must retain this copyright
-# notice, and modified files must contain a notice indicating that they have
-# been altered from the originals.
-#
-# Covalent is distributed in the hope that it will be useful, but WITHOUT
-# ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
-# FITNESS FOR A PARTICULAR PURPOSE. See the License for more details.
-#
-# Relief from the License may be granted by purchasing a commercial license.
-
-import covalent as ct
-from covalent._shared_files import logger
-from utils.executor_instance import executor
-
-app_log = logger.app_log
-log_stack_info = logger.log_stack_info
-
-
-@ct.electron(executor=executor)
-def join_words(a, b):
-    return ", ".join([a, b])
-
-
-@ct.electron(executor=executor)
-def excitement(a):
-    return f"{a}!"
-
-
-@ct.lattice
-def simple_workflow(a, b):
-    phrase = join_words(a, b)
-    return excitement(phrase)
-
-
-dispatch_id = ct.dispatch(simple_workflow)("Hello", "World")
-app_log.debug(f"AWS Lambda functional test `basic_workflow.py` dispatch id: {dispatch_id}")
-
-result = ct.get_result(dispatch_id, wait=True)
-status = str(result.status)
-
-assert status == str(ct.status.COMPLETED)
